Limit the rss feed to the most recent 50 items
Previously we'd put every single article in there which caused the feed size to grow pretty large. And now that we're including the full content of articles in the feed, its grown even more so to the point that we're hitting size limits on some feed validators.

At 50 items the size looks like it's back down to around 500kb.